Support for advanced payload autocompletion like query dsl (#124)
This PR adds the initial support for more advanced payload autocompletion for things like the Query DSL and Index Settings.

The spec is built by parsing and evaluating the TypeScript spec files from Kibana project (with Peek's own Parser and VM!). The process is rough and hacky. There are likely missed things and incorrect parsing. But it can already provide many desired features like query dsl completion. It is in a state that is sufficient to close #63. Further improvements should be done with separate and smaller issues.

The JS spec is merged with the JSON spec for completion. Also separate the API completion from the main Completer.

Resolves: #63