Add parser events and listener (#135)
The parser now publishes events when building language constructs. Listeners are able
to learn the actual state the parser is at on the fly. Even when the whole parsing fails, 
e.g. if parsing of API call fails because payload is of invalid format, listeners will still be 
able to know that the last good state is at payload parsing. The planed usage of this
feature is for the completer, which needs partial parsing results.

Resolves: #129